security: Complete MEDIUM and LOW priority audit remediation (59 issues)
MEDIUM priority fixes (M-1 through M-22):
- M-1: Error message sanitization to prevent internal state leakage
- M-2: JWT key rotation with 1-hour graceful window
- M-3: WebSocket ping timeout (30s timeout, 15s ping interval)
- M-4: Request body size limit (1MB default)
- M-8: Connection state cleanup on disconnect
- M-11: NaN/Infinity validation in difficulty calculations
- M-12: Integer arithmetic (basis points) for satoshi calculations
- M-14: JWT issuer/audience claims validation
- M-15: Config validation with range checks
- M-16: Log sanitization utility module

LOW priority fixes (L-1 through L-37):
- Replace unwrap() with expect() containing context messages
- Fix manual div_ceil implementations
- Fix manual abs_diff patterns
- Replace manual Default impls with derive macros
- Fix unnecessary borrows in hasher calls
- Fix is_none_or clippy suggestions
- Add type_complexity allows where needed
- Remove unnecessary casts

All 1132 tests pass, clippy clean, release build verified.

+++ crates/ghost-keys/src/derivation.rs
@@ -150,7 +150,7 @@ fn constant_time_sub_if_gte(result: &mut [u8; 32], n: &[u8; 32]) -> u8 {
 
     // Constant-time select: result = (use_temp) ? temp : result
     for i in 0..32 {
-        let mask = (use_temp.wrapping_neg()) as u8; // 0xFF if use_temp==1, 0x00 if use_temp==0
+        let mask = use_temp.wrapping_neg(); // 0xFF if use_temp==1, 0x00 if use_temp==0
         result[i] = (temp[i] & mask) | (result[i] & !mask);
     }
